edit cabin functionality with image update added to function

+++ src/services/apiCabins.js
@@ -11,39 +11,50 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  //1. Create cabin
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-    "/",
-    ""
-  );
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imageName = hasImagePath
+    ? newCabin.image
+    : `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "");
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert({ ...newCabin, image: imagePath });
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  //1. Create/edit cabin
+
+  let query = supabase.from("cabins");
+
+  //A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+  //B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error("Cabin could not be created");
   }
 
-  //2.Upload image //https://snxksbzdaykibwprwxow.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+  //2.Upload image //https://snxksbzdaykibwprwxow.supabase.co/storage/v1/object/public/cabin-images/...
 
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  //3.Delete cabin if error uploading
+    //3.Delete cabin if error uploading
 
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
-    console.error(storageError);
-    throw new Error(
-      "Cabin image could not be uploaded and cabin was not created"
-    );
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
+      console.error(storageError);
+      throw new Error(
+        "Cabin image could not be uploaded and cabin was not created"
+      );
+    }
   }
+
   return data;
 }
